Added HUGE update
I added the ability to just enter a URL along with a username and password, so that I (or whomever) wouldn't have to manually enter the Spanish to English translations for the specific lesson/unit into a Python dictionary. There are still some issues, and the code is horrible. I am going to be working on making the code better soon.